refactor: 대시보드 관련 .jsx/.css 파일 경로 이동 및 구조 정리
